+++ srcs/back/game-service/asset/package.json
@@ -20,6 +20,8 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
+        "@fastify/cookie": "^11.0.2",
+        "@fastify/websocket": "^11.0.2",
         "@prisma/client": "^6.6.0",
         "@types/node": "^22.14.1",
         "axios": "^1.9.0",
